+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- feature: New conversation modal [#16](https://github.com/admineclipta/chadbot-frontend/pull/16)
+
 ## [Release]
 
 ### v2.5.0